+++ src/ray/gcs/gcs_server/gcs_resource_scheduler.h
@@ -126,7 +126,7 @@ class GcsResourceScheduler {
   ///
   /// \param required_resources The resources to be scheduled.
   /// \return The Sorted resources.
-  std::vector<size_t> SortRequiredResources(
+  std::vector<int> SortRequiredResources(
       const std::vector<ResourceSet> &required_resources);
 
   /// Schedule resources according to `STRICT_SPREAD` strategy.
